refactor: extracted record controller logic into helper methods.
